refactor(tasks): move motion loader and reset owners [skip ci]
