+++ src/unilab/tasks/motion_tracking/common/domain_randomization.py
@@ -22,7 +22,8 @@ from unilab.dr.dr_utils import (
 )
 from unilab.dr.types import RESET_TERM_GEOM_FRICTION, ResetRandomizationPayload
 from unilab.dtype_config import get_global_dtype
-from unilab.envs.motion_tracking.common.reset import build_motion_reference_state
+
+from .reset import build_motion_reference_state
 
 
 class MotionTrackingDomainRandomizationProvider(DomainRandomizationProvider):
