+++ src/unilab/tasks/motion_tracking/g1/motion_box_loader.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass
 
 import numpy as np
 
-from unilab.envs.motion_tracking.common.motion_loader import MotionData, MotionLoader
+from ..common.motion_loader import MotionData, MotionLoader
 
 
 @dataclass
